Added the windows local profile application data directory as one of the storage directory options and added it as the first choice if no other exist on start up.

+++ src/nl/mpi/arbil/LinorgSessionStorage.java
@@ -85,26 +85,32 @@ public class LinorgSessionStorage {
         System.out.println("cacheDirExists: " + cacheDirExists());
     }
 
+    // Move the storeage directory and change the local corpus tree links to the new directory.
+    // After completion the application will be closed!
     public void changeStorageDirectory(String preferedDirectory) {
-        boolean success = new File(storageDirectory).renameTo(new File(preferedDirectory));
-        if (!success) {
-            LinorgWindowManager.getSingleInstance().addMessageDialogToQueue("Could not move the storage directory.\n", null);
-        } else {
-            try {
-                Vector<String> locationsList = new Vector<String>();
-                for (ImdiTreeObject[] currentTreeArray : new ImdiTreeObject[][]{TreeHelper.getSingleInstance().remoteCorpusNodes, TreeHelper.getSingleInstance().localCorpusNodes, TreeHelper.getSingleInstance().localFileNodes, TreeHelper.getSingleInstance().favouriteNodes}) {
-                    for (ImdiTreeObject currentLocation : currentTreeArray) {
-                        locationsList.add(currentLocation.getUrlString().replace(storageDirectory, preferedDirectory));
+        if (JOptionPane.YES_OPTION == JOptionPane.showConfirmDialog(LinorgWindowManager.getSingleInstance().linorgFrame, "Arbil will need to close in order to change the working directory.\nDo you wish to continue?", "Arbil", JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE)) {
+            boolean success = new File(storageDirectory).renameTo(new File(preferedDirectory));
+            if (!success) {
+                LinorgWindowManager.getSingleInstance().addMessageDialogToQueue("Could not move the storage directory to the requested location.\n", null);
+            } else {
+                try {
+                    Vector<String> locationsList = new Vector<String>();
+                    for (ImdiTreeObject[] currentTreeArray : new ImdiTreeObject[][]{TreeHelper.getSingleInstance().remoteCorpusNodes, TreeHelper.getSingleInstance().localCorpusNodes, TreeHelper.getSingleInstance().localFileNodes, TreeHelper.getSingleInstance().favouriteNodes}) {
+                        for (ImdiTreeObject currentLocation : currentTreeArray) {
+                            String currentLocationString = URLDecoder.decode(currentLocation.getUrlString(), "UTF-8");
+                            locationsList.add(currentLocationString.replace(storageDirectory, preferedDirectory));
+                        }
                     }
-                }
-                storageDirectory = preferedDirectory;
-                LinorgSessionStorage.getSingleInstance().saveObject(locationsList, "locationsList");
-                System.out.println("updated locationsList");
-            } catch (Exception ex) {
-                GuiHelper.linorgBugCatcher.logError(ex);
+                    storageDirectory = preferedDirectory;
+                    LinorgSessionStorage.getSingleInstance().saveObject(locationsList, "locationsList");
+                    System.out.println("updated locationsList");
+                } catch (Exception ex) {
+                    GuiHelper.linorgBugCatcher.logError(ex);
 //            System.out.println("save locationsList exception: " + ex.getMessage());
+                }
+                TreeHelper.getSingleInstance().loadLocationsList();
+                System.exit(0); // TODO: this exit might be unrequired
             }
-            TreeHelper.getSingleInstance().loadLocationsList();
         }
     }
 
@@ -112,7 +118,9 @@ public class LinorgSessionStorage {
         return new String[]{
                     // System.getProperty("user.dir") is unreliable in the case of Vista and possibly others
                     //http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6519127
+                    System.getenv("APPDATA") + File.separatorChar + "Local Settings" + File.separatorChar + "Application Data" + File.separatorChar + ".arbil" + File.separatorChar,
                     System.getenv("APPDATA") + File.separatorChar + ".arbil" + File.separatorChar,
+                    //                    System.getProperty("user.home") + File.separatorChar + "directory with spaces" + File.separatorChar + ".arbil" + File.separatorChar,
                     System.getProperty("user.home") + File.separatorChar + ".arbil" + File.separatorChar,
                     System.getenv("USERPROFILE") + File.separatorChar + ".arbil" + File.separatorChar,
                     System.getProperty("user.dir") + File.separatorChar + ".arbil" + File.separatorChar,
